moved message parser from Message to RocketChatMessage
