+++ src/main/java/io/rocketchat/core/uploader/FileUploader.java
@@ -1,14 +1,8 @@
 package io.rocketchat.core.uploader;
 
 /**
- * Created by sachin on 17/8/17.
+ * Created by sachin on 18/8/17.
  */
+
 public class FileUploader {
-
-    public static final String STORAGE_TYPE_S3 = "AmazonS3";
-    public static final String STORAGE_TYPE_GRID_FS = "GridFS";
-    public static final String STORAGE_TYPE_FILE_SYSTEM = "FileSystem";
-    public static final String STORAGE_TYPE_GOOGLE = "GoogleCloudStorage";
-
-
 }
