+++ src/main/java/io/rocketchat/core/uploader/RocketChatFile.java
@@ -3,6 +3,5 @@ package io.rocketchat.core.uploader;
 /**
  * Created by sachin on 18/8/17.
  */
-
-public class FileUploadImpl {
+public class RocketChatFile {
 }
